Support for generating broker keystore
This commit adds support to the Generate Configuration dialog for
creating a keystore for a new broker. The keystore includes a newly
generated broker certificate and key, client and broker CA bundles, and
a broker CA list file (if present).

Configuration settings for the brokerCaKeyFile, brokerCaPassword, and
brokerCaListFile were added to the Certificates section of the
dxlconsole config file. When not present, the locations of the
brokerCaKeyFile and brokerCaListFile are inferred from the location (if
specified) for the brokerCaBundleFile setting. brokerCaPassword defaults
to an empty password if not specified.